First commit
This commit is the repo setup, it contains the base of an Pong game.
This commit is meant the a base to build SPPT (Super Paper Pong
Tournament) upon.

As only decided to create a git repo now, there is a lot of undocumented
code here, I have the intention to document this code better and create
better commit from now on.