Changing render method to compatibility
As said by Juan Linietsky on twitter, for 2D games is stil better to use
the compatibility mode. Thats because the forward+ mode (vulkan) is
still compiling far more shaders than it should, what can be slow on
older computers.

As this is an simple project, that I don't intent to use nothing too
complex in terms of graphics, I think it is better to stick to
compatibility mode.