Random number generation fix
Before this commit, I was calling `randomize()` every time a Ball object was
instantiated, wich was every time the "Tela" scene was reloaded, wich
was every time a score was marked.

This was inneficient. The `randomize()` method affects the global scope,
so it is meant to be called only a single time, that's why I moved this call
to the `Game.gd` singleton. Now the `randomize()` method is called only a
single time when the game is launched and the `Game.gd` singleton is
loaded.

There are some other changes made to the `Ball.gd` file:
- I commented the line where I was defining direction fixed at 30
  degrees and now I'm defining it at random.
- On the "body_entered" signal functions, I fixed the type hint for the
  `body` variable, it should be `Node2D` instead of `Object`.