Some refactoring
Reloading the scene and deciding wich player serves shouldn't be dealt
by the Ball, thats why I renamed the "Tela" scene to "Match" and now it
is responsible for this.

This game is supposed to have PvP and PvE, on PvE scenrios, the oponent
is called "oponent", but in PvP you only have the "Player 1" and the
"Player 2". Thats why, now, I'm calling everything that's general about
the players using the names Player 1(2). I'm only using the name
"oponent" when referring specifically to the PvE AI.

Changed some immutable variables to constants, as it should have been
since the beggining.

It wasn't necessary to update the Player1(2)Score.text when a point was
scored because the scene would be reloaded and the text would be set at
the "_ready" function anyway.

Change some nomeclature to be more self explanatory:
- ball_initial_pos -> serving_player: the initial position of the ball
  is decided by wich player is serving.
- default_speed -> player_speed: this was confusing because we already
  have a variable called "ball_speed", so it wasn't clear what was this
referring to.
- local_shape_index == 0 -> local_shape_index == PLAYER_1(2)_AREA_IDX:
  using named constants instead of numbers is more self-explanatory.