Adding Multiplayer Scene
Created a multiplayer `Match` scene so that I can change modes from single
to multi. Created a second `Player` and adapted `Player.gd` with an
`is_player1` variable so that I am able to use diferent controls for
each player.

The variable `is_player1` is an exported variable and should be
modified on each specific instance, and not on the script itself.